Memory: Move LCD related areas to `Lcd`
I also moved `internal_memory` to `hardware`
