+++ emu/src/cpu/hardware/mod.rs
@@ -1,4 +1,5 @@
 pub mod dma;
+pub mod internal_memory;
 pub mod interrupt_control;
 pub mod keypad;
 pub mod lcd;
@@ -6,9 +7,19 @@ pub mod serial;
 pub mod sound;
 pub mod timers;
 
-pub trait HardwareComponent {
-    fn step(&mut self);
+pub const fn get_unmasked_address(
+    address: usize,
+    mask_get: usize,
+    mask_set: usize,
+    mask_shift: usize,
+    modulo: usize,
+) -> usize {
+    // Get the index of the mirror
+    let idx = (address & mask_get) >> mask_shift;
+    // Remove the mirror index from the address
+    let mut address = address & mask_set;
+    // Insert the unmasked index in the address
+    address |= (idx % modulo) << mask_shift;
 
-    // not sure if it will be useful, let's see
-    fn is_interrupt_pending(&self) -> bool;
+    address
 }
