+++ emu/src/lib.rs
@@ -3,5 +3,4 @@ pub mod bus;
 pub mod cartridge_header;
 mod cpu;
 pub mod gba;
-mod memory;
 pub mod render;
