--- emu/src/memory/mod.rs
@@ -1,2 +1,0 @@
-pub mod internal_memory;
-pub mod io_device;
